Ignore 'AllowedFast' messages when i metadata mode
These messages are only useful once we have the torrent
metadata and know the number of pieces in the actual torrent.
When in metadata mode the bitfield's length will correspond
to the number of pieces in the torrent metadata, so the odds
are high that an AllowedFast message will refer to an index
which does not exist anyway.