[Client] Accurately track protocol/data
Also we now never rate limit send invocations when sending
protocol. This ensures that protocol messages, like PieceRequests
are all sent in one go instead of being split into one or more
send requests.

This makes HTTPConnection a bit simpler as we don't have to deal
with partial sends where a RequestMessage is 1/2 in one buffer
and 1/2 in a second buffer