[dht] Check the TransactionId and ID to make sure they're valid
If we receive a message we should immediately ensure we have both
a transaction id and a valid ID. Some implementations seem to
be returning 320bit node IDs instead of 160bit node IDs. Maybe
they're hex encoding them (?!), but we should discard them
immediately instead of relying on errors being thrown later on
when the invalid ID is used.