[dht] Rewrite the initialise task to be more async
The task should be more optimal now as it sends FindNode requests
concurrently rather than sequentially. The worst-case performance
is still relatively poor (in the event the last messages all need
to hit four 15 second timeouts) so an overall cancellation/timeout
may be preferable. However, if we do that we may end up missing
the 'best' node when inserting ourselves into the DHT. This is
probably a fine tradeoff as we do re-query every few minutes and
re-insert in the new 'best' node.