[dht] Change the semantics of 'IDhtEngine.StartAsync'
The returned task used to wait for the full bootstrap to complete
(if one was needed). However that can take up to 2 minutes to
occur based on whether existing nodes could be restored, or the
utorrent router had to be contacted. In both cases we still
need to send a lot of Pings to other nodes, and any offline
nodes will be contacted 4 times at 15 seconds intervals before
they are marked as offline/non-contactable.

It is better to allow the returned task to be 'complete' as soon
as everything is set up. This allows the user to sanely cancel
the initialization by calling StopAsync at any point.

It also makes it possible to await this without having your app
feel like it's hanging due to the length network operations