[dht] GetPeers should keep track of active nodes only
When we search for peers, we should keep track of the nodes
which responded to us. This way when we are running the
Announce task we are more likely to annouce to the K closest
peers which are active, rather than the K closest peers of
which some (or all!) may be inactive.