[dht] Don't throw an exception when Nodes is missing
According to the spec a FindNodeResponse should contain a
'nodes' key, but some implementations are sending a response
with no nodes.

Let's give them the benefit of the doubt and assume that this
is because they have no nodes close to the value which was
queried and gracefully do nothing instead of throwing an
exception due to the missing key.