[dht] Just use a stopwatch directly for timeouting
If we want to time out a message early, we just set Engine.Timeout
to be Zero. We don't need to futz with the timing on the message
itself.

For everything else we can just call 'Seen' or 'Changed' and pass
in a large delta to make it appear like the node is old.