[core] Ensure ActivePeers and ConnectedPeers remain in sync
Always add/remove from these two lists at the same time to ensure
they are always in sync.

As part of these async-ify EndCheckEncryption so that we can
await it. This makes the logic a little clearer as we *do*
want encryption to be finalised before progressing with
any message sending/receiving.

Modified patchset from @OneFingerCodingWarrior