[core] Run PeerExchangeManager without using a timed callback
If we have a timer callback which executes every 1 minute, then
the objects referred to by that callback will not be eligible
for garbage collection until we return false to cancel the
re-invocation.

By running the PeerExchange logic without using a mainloop callback,
then we allow the objects to be garbage collected immediately after
we dispose the PeerId.

This means that when the PeerId is disposed we don't have to wait
up to 1 minute for the Timer callback to be invoked, so we can
return false, which means the delegate will be removed from the
Timer, which means the objects are (finally!) eligible for
garbage collection.

Partial fix for https://github.com/mono/monotorrent/issues/102.
This isn't a true leak, it's just slower-than-expected memory
release.