[core] Ensure we process every PeerId when switching to Seeding mode
If we're swapping from downloading -> seeding we dispose every
connection where we know the other peer is a seeder. We know we
will never need to send them data, or request data from them.

Every time we dispose a connection we should decrement the counter
so we don't skip an object by accident!

Partial fix for https://github.com/mono/monotorrent/issues/102.
This isn't a leak, it will just take a few more seconds than
expected to process the list of connections to remove all the
seeders.