[core] Small tweak after commit 3d54e0c7
Now that we're null out the Socket object we need to be careful
to null check the socket to avoid throwing a NullReferenceException
instead of an ObjectDisposedException.

It's 'normal' for us to dispose the socket while sending and
receiving simultaenously, so it's possible for the send or
receive to see a null socket.

This change just makes it a bit easier to debug while using
the debugger to suspend execution when an exception is thrown.