[core] Put the announce/scrape events on TrackerManager
This is easier than having to iterate over every Tracker and
attaching to individual events. As the Trackers themselves now
expose an 'AnnounceAsync' method we no longer need to follow the
pattern of using an event to track the completion of the async
operation. The 'Task' does that job for us now!

Also, Tracker, HTTPTracker and UdpTracker are implementation details
of the library. They should all be publicly exposed as `ITracker`.
The specific implementations don't have any type specific properties
which someone would need to investigate, and it is easy to disambiguate
between UDP/HTTP tracker by examining the Uri of the tracker (if you
ever need to do that for any reason).