[tests] Add a test for Scrapes throwing exceptions.
This also fixes two bugs :) Firstly, it is allowed to set
'Tracker' to null when all Trackers fail when we try to
Announce.

Secondly, we don't have to set the Tracker to null when we
fail to Scrape as we only scrape one tracker, we don't
do a round-robin and scrape them all.