[core] Simplify LocalPeerListener
It's concern is purely to receive broadcast messages and then
pass the data to the ClientEngine for it to be processed.

The ClientEngine is responsible for finding the correct
TorrentManager (in a thread-safe way) and adding the peer
to it.

Cancellation is handled via a normal CancellationTokenSource
and the UdpClient is disposed of whenever the CTS is cancelled.