[dht] Move IDhtEngine to the Dht namespace
The original reason for this split is that OpenSUSE wanted to ship
a build of MonoTorrent which did not contain any trace of a DHT
implementation and the code was pulled into a separate assembly
to give that guarantee.

The DHT interface lived in the main assembly to allow it to
integrate cleanly with the engine.