[core] Update the MagnetLink support
The MagnetLink object is immutable. There's a static Parse method
which takes the raw string, and a FromUri method which takes a Uri.

The constructor for MagnetLink allows you to construct one from
code (woo!) by passing in the various values. Now you don't need
to construct a magnet uri string in order to create a MagnetLink
object!

Good stuff.