[core] Try to incrementally hash pieces as they arrive
In the majority case we request blocks (more or less) in order
and so we can hash each block as it arrives. This means we will
probably have the hash ready by the time the final block arrives.

This is not 100% guaranteed as the PiecePicker does have logic to
allow multiple peers to download the same Piece. This logic was
tweaked to allow peers to prefer downloading their own piece, and
then to fall back to sharing a piece if there are no free pieces
available in the current 'rarest piece' subset.

In other words if the rarest set of pieces is calculated to be
[1, 2, 3] then the peer will choose one which has not been
started yet. If they have all been started, then one of them
will be helped. The helped piece will (nearly definitely) not
have a full incremental hash.

In local testing this seems to allow about 80% of pieces to
be fully incrementally hashed.