[core] Ensure we don't double network-convert the connection id
The connection ID should be written in network byte order, which
is handled automatically by the Message.ReadLong/WriteLong
methods.

Add a check on the server to make sure the value is correctly
received. Removing the extra HostToNetwork call makes the tests
green again, now that the extra sanity check is on the server
side.