[core] Tweak the semantics of TorrentManager.StopAsync
The task returned by this method will transition to the
'Completed' state only when the Mode has been set to
'Stopped'. This gives users an easy way to continue
execution after everything has stopped.

NOTE: This can take up to 1 minute if trackers need to be
contacted and they are all non-responsive. The typical
case should complete within a second or so.