[core] Ensure an IConnection is always passed to the PeerId constructor
Two different connections to the same ip/port/peerid should be treated
as distinct sessions as it could happen that a connection is closed,
and then a new one is created before the initial one has been fully
cleaned up.

The session is represented by the PeerId object, so PeerId objects
should not be compare as equal when they represent different
connections. As there's a 1:1 mapping between Connection and PeerId
the best option is to just remove the equality/gethashcode overrides.