[core] Optimise stopping a TorrentManager
As stopping a torrentmanager boils down to waiting for a few
Tasks to complete, we can simplify things.

StoppingMode now wraps up all the async tasks in a single Task
object which the TorrentManager can await. When that task
completes successfully we can immediately enter the Stopped
mode.

This makes things a little faster as we no longer need to wait
for the ClientEngine to run an iteration in order to progress
from the 'Stopping' mode to the 'Stopped' mode, we can safely
do it immediately after the tasks complete.