[core] Treat LocalPeerDiscovery like the Dht engine
The Dht engine is something we can Announce to, in order to inform
other peers we are downloading a torrent, and we can also get
notifications from the Dht engine when it finds peers we can use.

Local Peer Discovery is very similar. We can 'Announce' to it to
send the UDP broadcast message, and we can attach to it's PeerFound
event to be informed when peers are discovered.

This is now exposed under the simpler/unified ILocalPeerDiscovery
interface.