[core] Tweak DiskManager/PieceWriter to separate concerns
The DiskManager now stores/uses an IPieceWriter rather than
the old PieceWriter abstract class. To make this work the
helper methods which existed on the abstract class were moved
into DiskManager and combined with the existing methods to
keep things nice and simple.

DiskManager deals with TorrentManager instances and arrays of
TorrentFile instances, and PieceWriter deals exclusively with
TorrentFile objects.