[core] Make DiskManager get it's info from an interface
DiskManager now uses 'ITorrentData' to retrieve the list
of files, the torrent size and the piece length. This makes
it much easier to test the code as you don't have to create
an entire ClientEngine, TorrentManager and Torrent in order
to get the required data.

Also move the error handling code to the caller, not the
callee. If the torrentmanager (or whatever) calls a method
on DiskManager which throws an exception, it's up to the
caller to catch it and handle it. This removes the final
piece of code which required the DiskManager to be aware
of TorrentManagers.

As part of this 'DiskWriterTests' was renamed to
'DiskManagerExceptionTests' as it verifies the behaviour
when the DiskManager's PieceWriter throws an exception. A
new DiskManagerTests class was added which covers all the
other expected behaviours, like incremental hashing and
reading/writing data when a block spans multiple files.