[core] Test different combinations of incremental hashing
Add some tests verifying that writing the blocks for a Piece
with 3 blocks, in every permutation, results in a correct hash
using as much incremental data as possible.

The tests ensure that the expected amount of data is read from
the PieceWriter when data is written out-of-order. If the
DiskManager is backed by a MemoryWriter then any out-of-order
pieces have a high chance of being fulfilled from memory.