[tests] Don't use an actual Socket for the tests
Use a simple(ish?) implementation of a Stream that acts like
a socket instead. To mock a connection we create two instances
of the stream and swap the ends. That way each side of the
connection holds a Stream object they should read from, and one
they should write to.