[tests] Rewrite the encryption tests and EncryptorFactory
Rename TestEncryption.cs to EncryptorFactoryTests.cs and then redo
the tests to not rely on TorrentManager. As part of that expand
the set of tests to cover every combination of RC4Full, RC4Header,
PlainText and also initial payload.

EncryptorFactory is a little clearer now. The CheckIncoming method
will always return a non-null handshake, or it will throw an
exception. The CheckOutgoing method will allow you to append the
regular handshake so it is sent as part of the encrypted handshake.

The tests cover a PlainText connection connecting to an RC4Full only
peer, and also an RC4Full connection connecting to a PlainText only
peer. These are subtly different and are worth checking.

Fixed some minor issues in the encryption codepaths to cope with
this more thorough testing. Some exceptions weren't of the expected
type.

Did some small cleanup to EncryptedSocket and friends to lift
some variables from class-level to method-level. Also moved properties
which are specific to 'PeerAEncryption' or 'PeerBEncryption' out
of the base class and shared interface.