[tests] Remove duplicate tests and message send/receive helpers
TransferTest does not have to cover every possible encryption
combination as that's what EncryptorFactoryTests covers.

Remove a lot of the old helper methods to send/receive messages
as that can be handled by either NetworkIO (if you just want to
send/receive a guaranteed number of bytes), or PeerIO (if you
want to encrypt+send, or receive+decrypt) a full PeerMessage.