[core] Make PiecePicker.PickPiece Pass the Bitfield we choose from
If we change the semantics slightly we can make this a bit easier
to work with. PiecePicker.PickPiece should now pass the PeerId
we will pick the piece from, and also the Bitfield of pieces
which are available for us to choose.

It's the responsibility of PieceManager, or other PiecePickers,
to ensure that pieces which should not be chosen are removed
from the Bitfield before it gets to StandardPicker.

As part of this rewrite tests to better separate StandardPicker
tests from PieceManager tests.