[core] Move PiecePicker to use an IPieceRequestor instead of PeerId
PeerId is a massive grab-bag of everything imaginable. Attempt to
keep things separated, and easily testable, by using an IPieceRequestor
interface for piece picking which only exposes the properties
we *should* be using for piece picking.

For now PeerId implements the interface to reduce the number of
changes which are required for now. The interface itself may
change over the next few commits too.