[core] Allow SpeedMonitor to handle very fast speeds
Over a decade ago speeds in excess of 2GB/sec would've been
fairly unthinkable, but nowadays that's not too hard to
imagine.

As the SpeedMonitor is an averager, and since we're using a
fairly naive implementation which sums everything up and then
divides by the averaging period, we can easily overflow 32bits
if the rate gets above 150MB/sec as the default averaging period
is 12.

Rather than trying to squeeze into 32bits, just make all the
internals 64bit as it may become a requirement in the not too
distant future.