[core] Remove IRateLimiter.UpdateChunks from the interface
RateLimiterGroup should not expose an UpdateChunks method as
it is dangerous to do so. If there's a RateLimiter from the
ClientEngine and a RateLimiter for a TorrentManager in the
same group, we do not want to update the chunks for *both*
simultaenously.

Previously the TorrentManager was updating both of those
RateLimiters with it's own limit, effectively disabling the
overall engine limit.