[core] Add a PartialProgress public property
When some files are marked as DoNotDownload then this property
tracks the completion rate for the files which are marked as
downloadable.

If all downloadable files have been downloaded, the state will
change to 'Seeding' without telling the tracker the download has
completed.

If files which used to be marked as 'DoNotDownload' are then marked
as downloadable, the state will change back to Downloading.