[core] Move to IConnection2 to allow use of ReusableTask
This has been implemented in an API compatible way. Anyone
who has implemented IConnection will continue to be able
to use their implementation, and internally it will be
wrapped in an IConnection2 implementation.

That means all the built-in types can be moved to implement
IConnection2 directly, and so can be used directly.