[tests] Make this test reliable
The recent change to add the partial hash checking feature made this
test a little flakey. The mode was correctly detecting that the files
had not been fully hashed and was kicking off an asynchronous hash
check, thus causing the TorrentManager's bitfield to be set to
'false' as pieces failed their hashcheck. This meant that we would
not end up emitting the Announce request as expected.