[core] Immediately retry peers if handshaking fails
If we successfully connect to a peer, but then the handshake fails,
we should retry immediately.

If we originally tried an encrypted handshake, we should now try
an unencrypted one. If the remote peer doesn't support encryption
then this should give us a working connection immediately!

If we had tried a plaintext handshake first (because PreferEncryption
is set to false) then we should immediately retry an encrypted one.

If both encrypted and unencrypted fail, then we should not try to
connect to that peer again.