[core] Don't keep too many Peer instances alive
On active swarms we could discover hundreds or thousands of peers.
If we only connect to 60 of them we do not really want to store
thousands more in memory, so discard them.

If we ever fall below the maximum number of connections and do not
have sufficient peers to reach the maximum, we should quickly obtain
more via PEX, Dht or the Tracker.