[core] Rewrite the TorrentCreator to read and hash in parallel
This should give a bit of a speed boost. The use of 256kB
buffers for the file stream should minimize overheads. There's
very little to be gained by multi-threading reading from disk,
even on an SSD, as the best performance is always from
sequential reading.

Add in some timers (temporarily?) to help tune performance.