[core] Calculate the 'best' number of threads to use
If there are not many pieces in the torrent, don't use threads.
At the very least each thread we spin up should process 1 piece.
However if there are only a few pieces then it's just as easy to
create the torrent using a single thread.

I'm not aiming at the ultimate in hashing performance, i'm just
trying to get 'good enough' usage of the available CPU cores
and disk IO.