[core] Connection timeouts should not be debugger dependent
If execution is being paused in the debugger then connections
are going to be closed regardless of what we do.

Conversely, if the debugger is attached we should not make it
take aaaggees to connect to other peers because we disable the
timeout unnecessarily. OS default timeouts can be bad.