[core] Update the webseed tests a little
This ensures that HTTP connections are not accidentally
created for all the other tests.

TorrentManagers can have their logic executed without
running their 'StartAsync' method inside the test suite. So
in this scenario the Start time would not be set, which means
the http seeds would kick in immediately instead of being
delayed by 1 minute.

There are other ways to fix this, but this is a good start :p