[core] Change how EndGamePicker is used
In normal mode we want to pick a random piece from the
list of rarest pieces. We additionally take some other
factors into account, such as the relative priority of
different files.

When we enter end-game mode then we don't need any of
this complex logic anymore, all we want to do is request
any block we do not already have.

By swapping the order in which the PiecePickers are
instantiated we can get this behaviour more easily. The
extra benefit from the new ordering is that the check
to see if we should enter endgame mode will not be run
when we're trying to request a piece from a *subset* of
the available pieces, by virtue of how RarestPicker works.

Now we will only run the endgame check if all the complex
logic in RarestFirstPicker, RandomisedPicker, StandardPicker
etc results in no piece being selected from a peer which is
interesting.

I increased the threshold so that we can (theoretically)
enter endgame mode any time within the last 4MB of a given
torrent. The previous limit was too restrictive and only
allowed entering endgame mode within the last 320kB, which
is not great!