[core] Ensure DiskManager is not used after stopping
The aim behind this patch is to ensure asynchronous tasks
are properly ignored/cancelled when a torrent is stopped.

Part of our 'StoppingMode' closes all the open file streams
associated with a given torrent. This patch should ensure
that any pending 'WritePieceAsync' calls are discarded and
we do not actually re-open the stream we just closed.

In addition the send queue for active peers is cleared to
ensure that no 'PieceMessage' is accidentally processed
in between the time we Dispose a peer as part of shutdown
and when the 'ProcessQueue' method bails out due to the
connection closing.

Addresses part of https://github.com/mono/monotorrent/issues/190