[core] Ensure AmRequestingPiecesCount is accurate in endgame mode
When the final block in a given piece has been received we need
to remove the request from the list before attempting to cancel
other pending requests for that block.

This way we will not double-count that last block and decrement
AmRequestingPiecesCount by 2 instead of 1. This would have no
impact on the ability to complete a torrent, but it's better to
be correct :p