[core] Ensure TorrentFile progress is always accurate
If a torrent is hash checked and a file is marked as
'do not download' then we skip that hash check. The
overall progress of the torrent will always be less
than 100% in this case as some files have not been
hashchecked, and so we cannot guarantee the data is
there and is correct.

While it's unlikely to occur in practice, we should
ensure that the status of individual TorrentFiles
is refreshed if a hashcheck is performed while the
file is marked as 'DoNotDownload'. This will ensure
files do not report as 100% because they *used* to
be 100%.

If the files are eventually marked as higher priority
then they will be hash checked and their status will
be accurately reported again.