[core] Sanity check the bitfield length when connecting
Before we allow a peer to be added to the 'Connected' list,
ensure it's bitfield is of the correct length.

If we were in metadata mode before handshaking occurred and
we're in regular mode after it occurs, we could be left
with a peer with an incorrect bitfield.

May address https://github.com/mono/monotorrent/issues/196