[core] Robust loading of invalid torrents
Like most BitTorrent BEPs, the BEP for DHT is incredibly vague.
The only client which I can find which creates a 'nodes' key
in the torrent is BitComet and it creates a BEncodedList.

The original fix https://github.com/Sonarr/Sonarr/commit/3bc6bf9e992b0c2e0f82591608f2037fdd313297
looks like it's trying to handle the case where the 'nodes'
key is a BEncodedString (or some other non-list type). As
MonoTorrent is expecting a BEncodedList, and as the BEP vaguely
describes this key as being a list of sorts, then let's treat it
as such and ensure we discard anything which is not a list.