[core] Add a timeout when stopping torrents
This allows a relatively simple override for the final announce
so that the returned task doesn't hang around forever.

This is not implemented in the best of ways. Ideally the
ITracker and ITrackerManager interfaces would be augmented
to take a CancellationToken or TimeSpan parameter to handle
cancellation/timeouts, but that would be an ABI break which
I don't want to do right now.

This approach isn't great as it doesn't actually cancel the
pending request, but it's good enough and allows people to
use the timeout functionality in a way which will be
compatible with any future changes to ITracker/ITrackerManager.