[core] Add a way to stream media while downloading
This is a built in way to stream media while it is
actively being downloaded.

This feature supports seeking, and uses StreamingPiecePicker
to ensure data is downloaded on an as-needed basis.

There are two ways of consuming the data. Either a regular
Stream can be created using StreamProvider, or this can be
exposed over a basic HTTP endpoint so media players such as
VLC can play the stream.