[core] Retain peers when restarting torrents
We no longer clear the peer list when stopping
a torrent manager. Instead the peers are marked
as 'MaybeStale' the next time the manager is
started. We will connect to those as normal, but
if an announce occurs and we get new peers, we'll
replace any peers which are marked a 'MaybeStale'
with fresh ones which are more likely to be online.

Addresses https://github.com/alanmcgovern/monotorrent/issues/221