[core] Optimise memory for TorrentFile
Rather than storing one bitfield per TorrentFile just
to represent the start/end piece, we can use a simple
ValueTuple. The bitfield is used as a selector to
simplify some operations, but a ValueTuple representing
the (inclusive) start/end piece can perform the same
job.

The typical usecase is when some files are marked as
'DoNotDownload' and the selector is used to exclude
those pieces in the PiecePicker.