[core] Split BEncodedValue decoding out
The logic is all shared in one class now.
As part of this there was a mild refactoring
of the logic. A Decode/Encode loop on a 1MB
torrent file is 2.1 faster and allocates 30%
less memory. A net win for everyone!

A second benefit is that the Decode process
no longer needs to Peek at the stream in
order to decode the BEncoded data.