[core] Update bitfields faster when hashing
When we're hashing torrents with hundreds or thousands of files
we can update the bitfield more efficiently if we use binary
search to find a matching file.

In addition we can bail out as soon as we find a file which no
longer contains the piece we're interested in, which is a saving
in and of itself.

Partially addresses https://github.com/alanmcgovern/monotorrent/issues/223