[core] Ensure crypto is not on the engine thread
Push all the socket crypto work to a background thread.
While it's not a lot of CPU time, it's still stuff which
should be executed on the threadpool so the main loop is
available for syncing, and also to handle things like
StopAsync/StartAsync etc.

Partially addresses https://github.com/alanmcgovern/monotorrent/issues/223