[core] Add two states to the TrackerState enum
Add an 'Unknown' state to represent when we haven't
sent an announce or scrape request to a tracker yet.

Add a 'Connecting' state to represent when we're
sending an Announce or Scrape request.

NOTE: There's the potential for trampling here as
the Announce and Scrape codepaths both write to
the same 'FailureMessage', 'WarningMessage' and
'State' variables. Fixing this will be an API
break.

Fixes https://github.com/alanmcgovern/monotorrent/issues/255