[core] Make it easier to download torrent metadata
Add a method on ClientEngine which allows someone to
download the metadata for a torrent without creating
a TorrentManager and setting everything up themselves.

The TorrentManagers created behind the scenes to download
the metadata will not be visible to users. I'd love to
(eventually) refactor things so that a TorrentManager
never has to transition between 'MetadataMode' and
normal downloading/seeding mode. The ideal flow, for me,
would be for users to call the new DownloadMetadataAsync
method, and then once they receive that metadata they
can save it to disk and then allow the user to specify
which files they want to download, etc.

Fixes https://github.com/alanmcgovern/monotorrent/issues/242