[core] Relax the constraint that peerids should match
If the tracker gives us a peerid we should not *required*
the peer to have the exact same peerid. Some clients
deliberately randomize their ids (uTorrent), and others
may do so too.

uTorrent docs:

bt.randomize_peer_id: Disabling this option tells µTorrent
not to change peerid on every connection. It still uses
unique one for tracker, and should also be random limit.
This randomization is not being used for private torrents

My reading of this is that uTorrent uses a unique peerid
for every tracker it contacts, and it uses a random peerid
for every connection established *if* the corresponding
torrent is *not* private.