[core] Add the ability to wait for metadata to download
StreamProvider can only work when we know which file
needs to be streamed. Add a simple way for users to wait
for the metadata to be downloaded, taking into account the
existence of cached versions of the metadata.

This also exposes a 'Files' property so there is ready
access to the piece of information we need to get a
local or http stream.