[core] Correctly handle TorrentSettings.UploadSlots == 0
If the TorrentSettings has set the upload slots to 0 it
is supposed to mean 'unlimited'.

This changeset fixes a bug whereby it this would be
interpreted as meaning 'there are no slots available'.

There's a bit of cleanup as part of this commit which
doesn't make sense to do separately, so it is going
to be a little trickier to review than strictly necessary.

The added tests cover the usecase.

Fixes part of https://github.com/alanmcgovern/monotorrent/issues/258