[core] Build in port forwarding support
This support is built on top of Mono.Nat but can be swapped
to any other implementation if needed.

The public API is quite limited. There are two new methods
on ClientEngine which can be used to enable/disable port
forwarding. There are also two new properties, one which
tells you if port forwarding is enabled and one which gives
you the status of the forwarded ports.