[core] Clean up more code from the Unchoker (#281)
Greatly simplify the code in ChokeUnchokeManager. We now run a review every minimumTimeBetweenReviews even if we currently have enough free slots
to unchoke all currently interested peers.

The need to track firstCall and isDownloading was removed. firstCall is
unnecessary as we already track 'time since last review' and know when
we have never run a review. isDownloading is unnecessary as we used
it to trigger a review when the torrentmanager goes from downloading
to seeding - we handle that case by resetting the timer for the last
review so we execute a review on the next iteration.