[core] Make DiskManager and PieceWriter async
This is a relatively small change in the grand scheme
of things, but it allows for a few improvements.

Originally only one read or write operation could happen
at any point in time. Now this has been improved so that
each individual file is guaranteed to have a single
reader/writer, but multiple files can be accessed
concurrently.

The benefit out of this is that if some torrents are being
saved to slow storage, they will have zero impact on torrents
being downloaded to fast storage.