[core] Move to HttpClient as it's more modern
Also, the HttpTracker tests are flakey under .NET Core
as it appears to have a buggy implementation of
HttpWebRequest.Abort. The tests reliably reproduce a case
where you can call `Abort` on the request, and then the
subsequent call to 'GetResponseAsync' does not immediately
throw an exception like it's supposed to.