[core] Properly handle bitfield messages in metadata mode
We should ignore them when they're received. This is
implemented by special casing the situation where a
BitfieldMessage is received while we do not have
metadata - in this scenario we return the special value
'BitfieldMessage.UnknownValue' which we can check for
later.

If we attempt to *send* this to a peer it will error
out when you invoke 'Encode'. The MetadataMode overload
for 'HandleBitfield' asserts that all bitfields were
decoded as this sentinal value.

Before this patch we could only retrieve metadata from
peers which have no data, or from seeders. With this
patch we can download metadata from leechers too.

Fixes https://github.com/alanmcgovern/monotorrent/issues/302