[core] More accurately track PeerIds in the Connected list
As soon as we create a PeerId we put it in the Connected list.
This means that if we are in metadata mode and try to swap
to full downloading mode, we can close connections which were
created using the 'wrong' bitfield.

This closes a gap where we would *not* close the connection
if the mode switch happened while we were sending the handshake
message.

Addresses https://github.com/alanmcgovern/monotorrent/issues/304