[tests] Update this test to be more correct
We cannot have proper piece picking in metadata mode as we do
not acutally know how many pieces the torrent has. This test
had begun to fail as other parts of the codebase handle this
case more correctly (enforcing the invariant). As such it's
time to update this test so it uses a proper torrent.