[core] Wrap buffers in an IDisposable to simplify lifespan
Now we can wrap buffer usage in a 'using' statement in the
majority of cases, which is a neater way of representing
their lifespan.

In other cases we store the IDisposable and use that to return
the buffer to the pool when we're ready.

This should make it hard to accidentally double-free a buffer.