[core] Store the SocketAsyncEventArgs with the byte[]
Make BufferPool return an object which allows us to
cache the SocketAsyncEventArgs alongside the actual
byte[].

In addition the engine now creates two pools of buffers.
One is used for disk io, one is used for socket io. This
ensures that the byte[] we attach to SocketAsyncEventArgs
are always resused for further network operations. This
is a small performance win as this creating this is not
free.