[core] Emit the MetadataReceived event at the right time
This should happen after we set the Torrent metadata.
The original use for this event was just to allow
ClientEngine.DownloadMetadataAsync to download/save
the .torrent metadata.

I incorrectly reused it for StreamProvider to detect when
the metadata has been fully downloaded *and* associated
with the owning TorrentManager.

This patch changes the semantics of the event so that
it really does get invoked after the metadata is
downloaded and associated.