[dht] Remove an unnecessary DhtEngine.Add call
When we send a FindNode message to a torrent, we do
not want to add *every* node to the DhtTable. We
expect to discover thousands, or more, of nodes
as we search for the closest node to our ID.

As we search for the right location to announce
ourselves, we add nodes to the table by virtue of
them responding to a query we send them. This
means we don't need a separate 'ping' request.

By removing this call, we don't send Ping messages
to every single one of these 1000s of peers when
we only want to query a few dozen.