[dht] Allow the Initialization task to complete early
We now allow the 'Initialization' task to indicate it has
completed before it has fully inserted it's ID into the
DHT table.

This is to allow the engine to report as 'ready' for
querying as soon as there are a sufficient number of
active nodes in the table which can be queried.

This allows the DHT engine to report as 'Ready' within
a few seconds. Before this change all pending FindNode
queries would have to complete, and due to the nature
of how timeouts are handled, this could result in an
unbounded amount of time before the table was considered
healthy.