[dht] End GetPeersTask when we find enough peers
Add an early exit to this task. If we find at least 128 peers
we should stop sending further queries as we have probably
discovered more than enough peers.

The pending queries will still return more peers, so the 128
is not a hard limit, we may get dozens, or hundreds, more than
this.