[core] Compute the correct infohash for misordered dictionaries
If a BEncodedDictionary has been encoded with the keys in the
wrong order, we will now compute the infohash based on the
underlying bytes. This is the other spec compliant approach.

The first spec compliant approach is to reject the torrent.
Unfortunately there are plenty of invalid torrents out there,
so at this stage we may as well do the work to support this.

This should not be any worse performance wise as the original
approach.