[core] Fix a threading issue in FlushAsync
We should ensure we're executing on the IO thread
when we're interacting with the filestreams.

The 'Tick' method is invoked from the main engine's
tick so it'll be on the wrong thread unless we swap
here.

Also, ensure we take an async lock on the filestream
we intend to flush. This ensures there's multi-threaded
IO, but each stream is accessed in a single-threaded way.

Fixes https://github.com/alanmcgovern/monotorrent/issues/345