[core] Consolidate the 'Rented' and 'Locker' properties
The filestreams we use to read/write data from disk are now
protected with just a SemaphoreSlim.

The async lock is taken in the same place we used to set the
'Rented' property to true, and completely replaces that 'Rented'
boolean.

This allows us to move the 'Locker' from public API to private
API, hiding the details of making disk io threadsafe.

Fixes https://github.com/alanmcgovern/monotorrent/issues/348